services/zimmer-13-api-coordinator/src/services/credential_manager.py
"""
Credential Manager Service
Handles encrypted credential storage, retrieval, and rotation
"""

from sqlalchemy import Column, String, DateTime, Text, Boolean, Integer, JSON
from sqlalchemy.orm import Session
from cryptography.fernet import Fernet
from datetime import datetime, timedelta
import os
import base64
from typing import Optional, List, Dict, Any
import logging

logger = logging.getLogger(__name__)

class CredentialManager:
    """Manages encrypted credential storage and lifecycle"""

    def __init__(self, db: Session, encryption_key: Optional[str] = None):
        self.db = db
        # Initialize cipher with encryption key from env or parameter
        key = encryption_key or os.getenv('ENCRYPTION_KEY', None)
        if not key:
            # Generate new key if not provided
            key = Fernet.generate_key().decode()
            logger.warning("ENCRYPTION_KEY not set, generated new key. Set env var for persistence!")
        self.cipher = Fernet(key.encode() if isinstance(key, str) else key)

    def encrypt_value(self, value: str) -> str:
        """Encrypt credential value using Fernet (AES-128)"""
        try:
            encrypted = self.cipher.encrypt(value.encode())
            return base64.b64encode(encrypted).decode()
        except Exception as e:
            logger.error(f"Encryption failed: {e}")
            raise

    def decrypt_value(self, encrypted_value: str) -> str:
        """Decrypt credential value"""
        try:
            encrypted = base64.b64decode(encrypted_value.encode())
            decrypted = self.cipher.decrypt(encrypted)
            return decrypted.decode()
        except Exception as e:
            logger.error(f"Decryption failed: {e}")
            raise

    def create_credential(self, 
                         name: str,
                         credential_type: str,
                         service_name: str,
                         value: str,
                         description: Optional[str] = None,
                         metadata: Optional[Dict[str, Any]] = None,
                         expires_at: Optional[datetime] = None) -> Dict[str, Any]:
        """Create new credential with encryption"""
        
        encrypted_value = self.encrypt_value(value)
        
        credential = {
            'id': name,  # Use name as ID for simplicity
            'name': name,
            'credential_type': credential_type,
            'service_name': service_name,
            'encrypted_value': encrypted_value,
            'value': None,  # Don't return plaintext
            'description': description,
            'metadata': metadata or {},
            'expires_at': expires_at,
            'created_at': datetime.utcnow(),
            'updated_at': datetime.utcnow(),
            'last_used': None,
            'rotation_required': False,
            'rotation_count': 0
        }
        
        logger.info(f"Created credential: {name} for service: {service_name}")
        return credential

    def get_credential(self, credential_id: str, decrypt: bool = True) -> Optional[Dict[str, Any]]:
        """Retrieve credential, optionally decrypted"""
        # In real implementation, query from database
        # For now, this is a placeholder
        logger.info(f"Retrieved credential: {credential_id}, decrypt: {decrypt}")
        return None

    def get_service_credentials(self, service_name: str) -> List[Dict[str, Any]]:
        """Get all credentials for a specific service"""
        logger.info(f"Retrieving all credentials for service: {service_name}")
        return []

    def update_credential(self, credential_id: str, **kwargs) -> Optional[Dict[str, Any]]:
        """Update credential"""
        logger.info(f"Updated credential: {credential_id}")
        return None

    def delete_credential(self, credential_id: str) -> bool:
        """Delete credential"""
        logger.info(f"Deleted credential: {credential_id}")
        return True

    def rotate_credential(self, credential_id: str, new_value: str) -> Optional[Dict[str, Any]]:
        """Rotate credential to new value"""
        encrypted_value = self.encrypt_value(new_value)
        logger.info(f"Rotated credential: {credential_id}")
        return None

    def check_expiration(self) -> List[str]:
        """Check for expired credentials"""
        expired = []
        logger.info(f"Checked expiration: {len(expired)} expired credentials found")
        return expired

    def audit_access(self, credential_id: str, service_name: str, action: str):
        """Log credential access for audit"""
        log_entry = {
            'credential_id': credential_id,
            'service_name': service_name,
            'action': action,
            'timestamp': datetime.utcnow()
        }
        logger.info(f"Audit log: {log_entry}")
        return log_entry
