requirements.txt
playwright>=1.41.0
playwright-stealth>=1.0.6
aiohttp>=3.9.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
pydantic[email]
httpx>=0.26.0
tenacity>=8.2.3
mistralai>=0.1.0
google-generativeai>=0.4.1
redis>=5.0.1
python-dotenv>=1.0.1
uvicorn>=0.27.0
fastapi>=0.109.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
Pillow>=10.0.0
requests>=2.31.0
google-genai>=0.3.0
grpcio>=1.60.0
opencv-python-headless
numpy
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy[asyncio]
asyncpg
python-levenshtein
mcp
anthropic
openai
beautifulsoup4
lxml
pandas
psutil
docker
# Enterprise Audio & Proxy Support
SpeechRecognition
pydub
aiofiles
alembic
fake-useragent
imagehash
ultralytics
rich

pytesseract
